update
리뷰 입력,삭제시 책의 reviewCount 증가/감소

+++ BookReview/src/main/java/com/example/BookReview/ReviewDTO.java
@@ -11,7 +11,7 @@ import lombok.*;
 public class ReviewDTO {
     long id;
     String name;
-    int grade;
+    float grade;
     int bookId;
     String content;
 }
